Reduce warnings (#82)
* Remove references to directxmath package...

...in thprac/thprac.vcxproj

I can't figure out how to restore the package with NuGet, and as far as I can tell, it's straight-up unused in thprac anyway. Since I can't build the project otherwise, I'm removing the dependency(?) on the package.

* Don't disable C26812

This warning isn't actually raised during the build, for reasons I don't completely understand. Maybe it was left over from an earlier version of thprac?

* Fix C4244 (conversion from double to float)

I imagine the use of `std::stod` was in error, though I can't be entirely certain.

Running msbuild now produces 3 warnings, down from 56.

* Fix C4018 (signed/unsigned mismatch)

Running msbuild now produces 1 warning, down from 3.

The final warning is LNK4098, which I don't fully understand at this point in time.

+++ thprac/src/thprac/thprac_launcher_cfg.cpp
@@ -290,7 +290,7 @@ void LauncherSettingSet(const char* name, std::string& valueIn)
             if (jsonVec[i].IsNumber()) {                                        \
                 val = jsonVec[i].GetFloat();                                    \
             } else if (jsonVec[i].IsString()) {                                 \
-                val = std::stod(jsonVec[i].GetString());                        \
+                val = std::stof(jsonVec[i].GetString());                        \
             } else {                                                            \
                 throw std::runtime_error(#col ": wrong format");                \
             }                                                                   \
@@ -1262,7 +1262,7 @@ private:
                 mCfgTheme.Set(Sus);
                 return;
             }
-            for (int i = 0; i < userThemes.size(); i++) {
+            for (size_t i = 0; i < userThemes.size(); i++) {
                 if (!strcmp(userThemes[i].utf8, cur_theme)) {
                     userThemeIdx = i;
                     configThemeExists = true;
@@ -1706,7 +1706,7 @@ private:
             }
         }
         if (themeIsUser && ImGui::BeginCombo("##themes_user", userThemes[userThemeIdx].utf8)) {
-            for (int i = 0; i < userThemes.size(); i++) {
+            for (size_t i = 0; i < userThemes.size(); i++) {
                 bool selected = i == userThemeIdx;
                 if (ImGui::Selectable(userThemes[i].utf8, selected)) {
                     userThemeIdx = i;
